# neurosentio-copilot-agent/reports/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NeuroSentio Copilot Agent",
    "description": "The standalone backend brain of NeuroSentio Daily Copilot. Manages user profiles, tasks, energy logs, and generates rule-based and LLM-powered daily plans. Designed for neurodivergent-friendly pacing.",
    "version": "0.5.0"
  },
  "paths": {
...
    "/reply/draft": {
      "post": {
        "tags": [
          "ReplyDrafter"
        ],
        "summary": "Generate reply draft options",
        "description": "Paste any message text and receive 3 neurodivergent-friendly reply options:\nshort, warm, and detailed. A boundary option is included when:\n- current_energy < 40\n- user_intent includes decline/delay/boundary/not available\n- include_boundary_option=true (default)\n\n**This endpoint NEVER sends messages.**\nAll drafts are stored locally. No external connections.",
        "operationId": "create_reply_draft_reply_draft_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplyDraftRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplyDraft"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/reply/drafts": {
      "get": {
        "tags": [
          "ReplyDrafter"
        ],
        "summary": "List reply drafts",
        "description": "Returns the current user's reply drafts.\nDeleted drafts are excluded by default \u2014 use `status=deleted` to retrieve them.",
        "operationId": "list_reply_drafts_reply_drafts_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by status. Deleted drafts are excluded by default.",
              "title": "Status"
            },
            "description": "Filter by status. Deleted drafts are excluded by default."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReplyDraftListItem"
                  },
                  "title": "Response List Reply Drafts Reply Drafts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/reply/drafts/{draft_id}": {
      "get": {
        "tags": [
          "ReplyDrafter"
        ],
        "summary": "Get a single reply draft",
        "description": "Returns one reply draft by ID. Returns 404 if not found or belongs to another user.",
        "operationId": "get_reply_draft_reply_drafts__draft_id__get",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Draft Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplyDraft"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "ReplyDrafter"
        ],
        "summary": "Update a reply draft",
        "description": "Update a draft: select an option, save an edited reply, or change status.\nValid statuses: drafted | edited | selected | archived | deleted\nThe status 'sent' is intentionally not allowed \u2014 this service never sends messages.",
        "operationId": "update_reply_draft_reply_drafts__draft_id__patch",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Draft Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplyDraftUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplyDraft"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ReplyDrafter"
        ],
        "summary": "Soft delete a reply draft",
        "description": "Soft deletes the draft by setting status='deleted'.\nThe record is preserved in the database but excluded from default list results.\nUse GET /reply/drafts?status=deleted to retrieve deleted drafts.",
        "operationId": "delete_reply_draft_reply_drafts__draft_id__delete",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Draft Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplyDraftDeleteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/reply/drafts/{draft_id}/original-message": {
      "delete": {
        "tags": [
          "ReplyDrafter"
        ],
        "summary": "Purge/redact the original message from a reply draft",
        "description": "Individually purge the original message stored inside a reply draft,\nreplacing it with the placeholder '[redacted]'.",
        "operationId": "redact_original_message_reply_drafts__draft_id__original_message_delete",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Draft Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplyDraft"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReplyDraft": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "source_type": {
            "type": "string",
            "title": "Source Type"
          },
          "original_message": {
            "type": "string",
            "title": "Original Message"
          },
          "message_sender": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message Sender"
          },
          "message_subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message Subject"
          },
          "message_channel": {
            "type": "string",
            "title": "Message Channel"
          },
          "user_intent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Intent"
          },
          "preferred_tone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Tone"
          },
          "context_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Note"
          },
          "draft_options": {
            "items": {
              "$ref": "#/components/schemas/ReplyDraftOption"
            },
            "type": "array",
            "title": "Draft Options"
          },
          "selected_option_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Option Type"
          },
          "edited_reply": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Edited Reply"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "source_type",
          "original_message",
          "message_channel",
          "draft_options",
          "status",
          "source",
          "created_at",
          "updated_at"
        ],
        "title": "ReplyDraft"
      },
      "ReplyDraftDeleteResponse": {
        "properties": {
          "deleted": {
            "type": "boolean",
            "title": "Deleted"
          },
          "id": {
            "type": "string",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "deleted",
          "id"
        ],
        "title": "ReplyDraftDeleteResponse"
      },
      "ReplyDraftListItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "message_sender": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message Sender"
          },
          "message_subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message Subject"
          },
          "message_channel": {
            "type": "string",
            "title": "Message Channel"
          },
          "user_intent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Intent"
          },
          "selected_option_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Option Type"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "message_channel",
          "status",
          "source",
          "created_at",
          "updated_at"
        ],
        "title": "ReplyDraftListItem"
      },
      "ReplyDraftOption": {
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "short",
              "warm",
              "detailed",
              "boundary"
            ],
            "title": "Type"
          },
          "text": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "type",
          "text"
        ],
        "title": "ReplyDraftOption"
      },
      "ReplyDraftRequest": {
        "properties": {
          "original_message": {
            "type": "string",
            "maxLength": 10000,
            "minLength": 3,
            "title": "Original Message"
          },
          "message_sender": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message Sender"
          },
          "message_subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message Subject"
          },
          "message_channel": {
            "type": "string",
            "enum": [
              "manual",
              "email",
              "sms",
              "whatsapp",
              "slack",
              "other"
            ],
            "title": "Message Channel",
            "default": "manual"
          },
          "user_intent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Intent"
          },
          "preferred_tone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Tone"
          },
          "context_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Note"
          },
          "include_boundary_option": {
            "type": "boolean",
            "title": "Include Boundary Option",
            "default": true
          },
          "max_length": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "short",
                  "medium",
                  "detailed"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Length"
          },
          "current_energy": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Energy"
          }
        },
        "type": "object",
        "required": [
          "original_message"
        ],
        "title": "ReplyDraftRequest"
      },
      "ReplyDraftUpdate": {
        "properties": {
          "selected_option_type": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "short",
                  "warm",
                  "detailed",
                  "boundary"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Option Type"
          },
          "edited_reply": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Edited Reply"
          },
          "status": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "drafted",
                  "edited",
                  "selected",
                  "archived",
                  "deleted"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "title": "ReplyDraftUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}